requirements.txt
python-dotenv
nixtlats
pandas
matplotlib